rotate model depending on current camera position

+++ src/de/swagner/mgcube/Block.java
@@ -2,9 +2,9 @@ package de.swagner.mgcube;
 
 public class Block {
 
-	public int x=0, y=0, z=0;
+	public float x=0, y=0, z=0;
 
-	public Block(int x, int y, int z) {
+	public Block(float x, float y, float z) {
 		this.x = x;
 		this.y = y;
 		this.z = z;
