+++ src/de/swagner/mgcube/Target.java
@@ -1,5 +1,5 @@
 package de.swagner.mgcube;
 
 public class Target {
-	public int x=0, y=0, z=0;
+	public float x=0, y=0, z=0;
 }
